refactor(calendar): extract busy-interval computation into services for reuse + MED-284
