+++ backend/calendars/views.py
@@ -19,6 +19,12 @@ from .services import (
     cancel_single_occurrence,
     split_series_from_occurrence,
     _count_occurrences_before,
+    # Occurrence expansion moved to services so booking availability can reuse
+    # it without importing the API layer. Re-exported here: existing callers
+    # (including tests) still import these names from views.
+    _events_intersecting_range,
+    _expand_recurring_event,
+    get_busy_intervals_by_calendar,
 )
 
 from core.models import ProjectMember
@@ -620,32 +626,6 @@ def _get_accessible_calendars(
     return qs.distinct()
 
 
-def _events_intersecting_range(start_dt, end_dt, base_qs=None):
-    """
-    Return events that may appear in [start_dt, end_dt).
-
-    Non-recurring events use wall-clock overlap. Recurring masters are included
-    when the series can still produce instances in the window (so split-born
-    series remain visible after their first occurrence day).
-    """
-    if base_qs is None:
-        base_qs = Event.objects.all()
-
-    non_recurring = Q(
-        is_recurring=False,
-        start_datetime__lt=end_dt,
-        end_datetime__gt=start_dt,
-    )
-    recurring = Q(
-        is_recurring=True,
-        recurrence_rule__isnull=False,
-        start_datetime__lt=end_dt,
-    ) & (
-        Q(recurrence_rule__until__isnull=True)
-        | Q(recurrence_rule__until__gt=start_dt)
-    )
-    return base_qs.filter(non_recurring | recurring)
-
 
 def _build_calendar_view_payload(
     user,
@@ -710,100 +690,6 @@ def _build_calendar_view_payload(
         "calendars": calendars_data,
     }
 
-
-def _expand_recurring_event(
-    event: Event,
-    time_min,
-    time_max,
-    max_results: int = 250,
-):
-    """
-    Expand a recurring event into concrete instances within [time_min, time_max).
-    Currently supports simple DAILY and WEEKLY patterns based on start_datetime.
-    """
-    if not event.is_recurring or not event.recurrence_rule_id:
-        return []
-
-    rule = event.recurrence_rule
-    frequency = rule.frequency
-    interval = max(int(rule.interval or 1), 1)
-
-    duration = event.end_datetime - event.start_datetime
-    instances: list[Any] = []
-
-    # Load exceptions for this event/rule within range
-    exceptions = RecurrenceException.objects.filter(
-        organization=event.organization,
-        recurrence_rule=rule,
-        original_event=event,
-        exception_date__gte=time_min,
-        exception_date__lt=time_max,
-    ).select_related("modified_event")
-    exceptions_by_date = {exc.exception_date: exc for exc in exceptions}
-
-    # Fast-forward to first occurrence that could intersect [time_min, time_max)
-    if frequency == "DAILY":
-        step = timezone.timedelta(days=interval)
-    elif frequency == "WEEKLY":
-        step = timezone.timedelta(weeks=interval)
-    else:
-        # For now only basic DAILY/WEEKLY patterns are supported in expansion.
-        return []
-
-    # Honor the series bounds so a capped/split series stops generating.
-    # `until` is treated as exclusive (strict-less): an occurrence exactly at
-    # `until` belongs to the next (split) series, never the capped master.
-    rule_until = rule.until
-    rule_count = rule.count
-
-    # Skip occurrences that end at or before time_min (first that can intersect
-    # the window has start > time_min - duration).
-    occurrence_index = _count_occurrences_before(
-        event.start_datetime, time_min - duration, rule
-    )
-    current = event.start_datetime + (step * occurrence_index)
-
-    if rule_count is not None and occurrence_index >= rule_count:
-        return []
-    if rule_until is not None and current >= rule_until:
-        return []
-
-    while current + duration <= time_max and len(instances) < max_results:
-        if rule_count is not None and occurrence_index >= rule_count:
-            break
-        if rule_until is not None and current >= rule_until:
-            break
-
-        # Check intersection with requested window
-        if current < time_max and (current + duration) > time_min:
-            exc = exceptions_by_date.get(current)
-            if exc:
-                if exc.is_cancelled:
-                    # Skip cancelled instance
-                    pass
-                else:
-                    # Use modified event instance
-                    instances.append(exc.modified_event)
-            else:
-                # Create a lightweight instance based on the master event
-                attrs = {}
-                for field in Event._meta.fields:
-                    name = field.name
-                    attrs[name] = getattr(event, name)
-
-                # Override fields specific to this occurrence
-                attrs["id"] = event.id  # master id; original_start differentiates instances
-                attrs["start_datetime"] = current
-                attrs["end_datetime"] = current + duration
-                attrs["original_start"] = current
-
-                instance_obj = SimpleNamespace(**attrs)
-                instances.append(instance_obj)
-
-        current = current + step
-        occurrence_index += 1
-
-    return instances
 
 
 class EventInstancesView(generics.ListAPIView):
@@ -1475,53 +1361,19 @@ class FreeBusyView(generics.GenericAPIView):
             "calendars": {},
         }
 
-        for cal in calendars:
-            events_qs = _events_intersecting_range(
-                time_min,
-                time_max,
-                Event.objects.filter(
-                    calendar=cal,
-                    is_deleted=False,
-                ).select_related("recurrence_rule"),
-            )
+        # Expansion + merge live in services so booking availability reuses the
+        # same computation rather than a second copy of it.
+        by_calendar = get_busy_intervals_by_calendar(calendars, time_min, time_max)
 
-            intervals = []
-            for ev in events_qs:
-                if ev.is_recurring and ev.recurrence_rule_id:
-                    instances = _expand_recurring_event(ev, time_min, time_max)
-                    for inst in instances:
-                        intervals.append(
-                            [
-                                getattr(inst, "start_datetime"),
-                                getattr(inst, "end_datetime"),
-                            ]
-                        )
-                else:
-                    intervals.append([ev.start_datetime, ev.end_datetime])
-
-            # Merge overlapping intervals
-            intervals = sorted(intervals, key=lambda x: x[0])
-            merged = []
-            for start, end in intervals:
-                if not merged:
-                    merged.append([start, end])
-                else:
-                    last_start, last_end = merged[-1]
-                    if start <= last_end:
-                        merged[-1][1] = max(last_end, end)
-                    else:
-                        merged.append([start, end])
-
-            busy = [
-                {
-                    "start": s.isoformat().replace("+00:00", "Z"),
-                    "end": e.isoformat().replace("+00:00", "Z"),
-                }
-                for s, e in merged
-            ]
-
-            result["calendars"][str(cal.id)] = {
-                "busy": busy,
+        for calendar_id, intervals in by_calendar.items():
+            result["calendars"][calendar_id] = {
+                "busy": [
+                    {
+                        "start": s.isoformat().replace("+00:00", "Z"),
+                        "end": e.isoformat().replace("+00:00", "Z"),
+                    }
+                    for s, e in intervals
+                ],
                 "errors": [],
             }
